LPD-48736 Merge DLStorageQuotaOSGiCommands.java and StoreAreaOSGiCommands.java into one class
